Omogućavanje dodavanje sastojaka receptu, fetchanje profila drugog korisnika

+++ IzvorniKod/CookBooked/src/main/java/opp/CookBooked/service/implementacija/KomentariServiceJpa.java
@@ -30,37 +30,60 @@ public class KomentariServiceJpa implements KomentariService {
 
     @Override
     public KomentariRecept dodajKomentar(Komentar komentar, long idRecept) {
-        Recept recept = receptRepo.findByIdRecept(idRecept);
-        KomentariRecept kr = new KomentariRecept();
-        komentar.setDatumKomentar(LocalDate.now());
-        kr.setKomentar(komentar);
-        komentariRepo.save(komentar);
-        kr.setRecept(recept);
-        komRecRepo.save(kr);
-        return kr;
+
+        try {
+
+            Recept recept = receptRepo.findByIdRecept(idRecept);
+            if (recept != null) {
+                KomentariRecept kr = new KomentariRecept();
+                komentar.setDatumKomentar(LocalDate.now());
+
+                kr.setKomentar(komentar);
+                komentariRepo.save(komentar);
+
+                kr.setRecept(recept);
+                komRecRepo.save(kr);
+
+                return kr;
+            } else return null;
+
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 
     @Override
     public String urediKomentar(long idKomentar, Komentar updatedKomentar) {
         Komentar kom = findById(idKomentar);
         Korisnik korisnik = korisnikRepo.findByIdKorisnik(kom.getKorisnik().getIdKorisnik());
-        kom.setOpisKomentar(updatedKomentar.getOpisKomentar());
-        kom.setDatumKomentar(LocalDate.now());
-        komentariRepo.save(kom);
-        return "Uspješno ažuriranje komentara korisnika " + korisnik.getKorisnickoIme();
+
+        if (kom != null && korisnik != null) {
+
+            kom.setOpisKomentar(updatedKomentar.getOpisKomentar());
+            kom.setDatumKomentar(LocalDate.now());
+            komentariRepo.save(kom);
+            return "Uspješno ažuriranje komentara korisnika " + korisnik.getKorisnickoIme();
+
+        } else return "Neuspješno";
     }
 
     @Override
     public String obrisiKomentar(long idKomentar, long idRecept) {
-        Komentar kom = findById(idKomentar);
-        Recept recept = receptRepo.findByIdRecept(idRecept);
-        KomentariRecept existingRelationship = komRecRepo
-                .findByReceptIdAndAutorIdAAndKomentarId(idRecept, kom.getIdKomentar(), kom.getKorisnik().getIdKorisnik());
-        if (existingRelationship != null) {
-            komRecRepo.delete(existingRelationship);
+        try {
+            Komentar kom = findById(idKomentar);
+            Recept recept = receptRepo.findByIdRecept(idRecept);
+            KomentariRecept existingRelationship = komRecRepo
+                    .findByReceptIdAndAutorIdAAndKomentarId(idRecept, kom.getIdKomentar(), kom.getKorisnik().getIdKorisnik());
+            if (existingRelationship != null && kom != null && recept != null) {
+                komRecRepo.delete(existingRelationship);
+            }
+            komentariRepo.delete(kom);
+            return "Uspješno brisanje komentara";
+        } catch (Exception e) {
+            e.printStackTrace();
+            return null;
         }
-        komentariRepo.delete(kom);
-        return "Uspješno brisanje komentara";
     }
 
     @Override
